Move the Store class to be inside transformation.
